categoryConverter

+++ src/main/java/com/ditsikts/tvseries/dao/TvSeriesDao.java
@@ -8,9 +8,9 @@ public interface TvSeriesDao {
 
 	public List<TvSeries> findAll();
 	
-	public TvSeries findById(int id);
+	public TvSeries findById(Long id);
 	
 	public void save(TvSeries tvSeries);
 	
-	public void deleteById(int id);
+	public void deleteById(Long id);
 }
